+++ src/Config/Async.php
@@ -5,14 +5,7 @@ namespace Innmind\Async\Config;
 
 use Innmind\OperatingSystem\Config;
 use Innmind\Signals\Async\Interceptor;
-use Innmind\Time\{
-    Clock,
-    Period,
-    Halt,
-};
-use Innmind\HttpTransport\Transport;
-use Innmind\IO\IO;
-use Innmind\Signals\Handler;
+use Innmind\Time\Clock;
 
 /**
  * @internal
@@ -30,29 +23,17 @@ final class Async
 
     public function __invoke(Config $config): Config
     {
-        $halt = Halt::async($this->clock);
-        $io = IO::async(
-            $config->io(),
-            $this->clock,
-        );
-        // todo build a native client based on innmind/io to better integrate in
-        // this system.
-        $http = Transport::async(
-            $this->clock,
-            $io,
-            Period::millisecond(10), // this is blocking the active task so it needs to be low
-            static fn() => $halt(Period::millisecond(1))->unwrap(), // this allows to jump between tasks
-        );
-        $signals = Handler::async(
-            $config->signalsHandler(),
-            $this->interceptor,
-        );
-
         return $config
-            ->haltProcessVia($halt)
-            ->useHttpTransport($http)
-            ->withIO($io)
-            ->handleSignalsVia($signals);
+            ->mapIO(fn($io) => $io->asAsync($this->clock))
+            ->mapHalt(fn($halt) => $halt->asAsync($this->clock))
+            ->mapHttpTransport(fn($transport, $config) => $transport->map(
+                fn($http) => $http->asAsync(
+                    $this->clock,
+                    $config->halt(),
+                    $config->io(),
+                ),
+            ))
+            ->mapSignalsHandler(fn($signals) => $signals->asAsync($this->interceptor));
     }
 
     /**